+++ src/pages/HomePage/BoxCategory/BoxCategory.tsx
@@ -15,7 +15,7 @@ import { Task } from '@/shared/types/home';
 import ButtonAddIcon from '@/shared/assets/svgs/btn_task_add.svg?react';
 import MeatBallDefault from '@/shared/assets/svgs/todo_meatball_default.svg?react';
 
-import { useCalendar } from '../../hooks/useCalendar';
+import { useCalendar } from '../hooks/useCalendar';
 import BoxTodoInput from './BoxTodoInput/BoxTodoInput';
 import StatusDefaultBoxCategory from './StatusDefaultBoxCategory/StatusDefaultBoxCategory';
 import { useCreateTodo } from './hooks/useCreateTodo';
